+++ vendor-prefixed/guzzlehttp/guzzle/src/PrepareBodyMiddleware.php
@@ -8,7 +8,7 @@
 namespace WPGraphQL\Login\Vendor\GuzzleHttp;
 
 use WPGraphQL\Login\Vendor\GuzzleHttp\Promise\PromiseInterface;
-use Psr\Http\Message\RequestInterface;
+use WPGraphQL\Login\Vendor\Psr\Http\Message\RequestInterface;
 
 /**
  * Prepares requests that contain a body, adding the Content-Length,