+++ vendor-prefixed/guzzlehttp/psr7/src/ServerRequest.php
@@ -10,10 +10,10 @@ declare(strict_types=1);
 namespace WPGraphQL\Login\Vendor\GuzzleHttp\Psr7;
 
 use InvalidArgumentException;
-use Psr\Http\Message\ServerRequestInterface;
-use Psr\Http\Message\StreamInterface;
-use Psr\Http\Message\UploadedFileInterface;
-use Psr\Http\Message\UriInterface;
+use WPGraphQL\Login\Vendor\Psr\Http\Message\ServerRequestInterface;
+use WPGraphQL\Login\Vendor\Psr\Http\Message\StreamInterface;
+use WPGraphQL\Login\Vendor\Psr\Http\Message\UploadedFileInterface;
+use WPGraphQL\Login\Vendor\Psr\Http\Message\UriInterface;
 
 /**
  * Server-side HTTP request