+++ vendor-prefixed/league/oauth2-google/src/Provider/Google.php
@@ -11,7 +11,7 @@ use WPGraphQL\Login\Vendor\League\OAuth2\Client\Exception\HostedDomainException;
 use WPGraphQL\Login\Vendor\League\OAuth2\Client\Provider\Exception\IdentityProviderException;
 use WPGraphQL\Login\Vendor\League\OAuth2\Client\Token\AccessToken;
 use WPGraphQL\Login\Vendor\League\OAuth2\Client\Tool\BearerAuthorizationTrait;
-use Psr\Http\Message\ResponseInterface;
+use WPGraphQL\Login\Vendor\Psr\Http\Message\ResponseInterface;
 
 class Google extends AbstractProvider
 {