+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to a modified version of [Semantic Versioning](./README
 ## Unreleased
 
 - fix: Avoid strict-typing `League\OAuth2\Client\Provider\AbstractProvider` to the `\WPGraphQL\Login\Vendor` namespace. H/t @pat-flew .
+- chore!: Use Strauss to prefix `Psr` dependencies. This is a breaking change, as the `Psr` dependencies are now prefixed with `\WPGraphQL\Login\Vendor`.
 - chore!: Bump minimum supported WordPress version to 6.0.
 - chore!: Bump minimum supported WPGraphQL version to 1.14.0.
 - chore: Update `WPGraphQL Coding Standards` to latest version and lint.